+++ lib/models/ice_cream_model.dart
@@ -1,5 +1,15 @@
 import '../entities.dart';
 
+/// [IceCreamModel] is a simple model class representing an ice cream with a flavor and size.
+///
+/// Example:
+/// ```dart
+/// final IceCreamModel vanillaIceCream = IceCreamModel('Vanilla', 'Large');
+/// print(vanillaIceCream);
+///
+/// final IceCreamModel chocolateIceCream = IceCreamModel('Chocolate', 'Medium');
+/// print(chocolateIceCream);
+/// ```.
 class IceCreamModel implements Model {
   const IceCreamModel(
     this.flavor,