+++ test/invokers/waiter_invoker_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:pc_command/commands/order_ice_cream_command.dart';
 import 'package:pc_command/invokers/waiter_invoker.dart';
 import 'package:pc_command/models/ice_cream_model.dart';
-import 'package:pc_command/receivers/ice_cream_receiver.dart';
+import 'package:pc_command/receivers/ice_cream_shop_receiver.dart';
 
 void main() {
   test('WaiterInvoker should execute and undo commands correctly', () {